Remove unsed objects
