+++ Source/nHydrate.Generator/INHydrateModelObject.cs
@@ -6,6 +6,6 @@ namespace nHydrate.Generator
     {
         INHydrateModelObject Root { get; }
         string Key { get; }
-        INHydrateModelObjectController Controller { get; set; }
+        //INHydrateModelObjectController Controller { get; set; }
     }
 }
